backend/tweets/requirements.txt
requests
pandas
python-dotenv>=1.0.0
matplotlib
twitter-api-client
python-dotenv
selenium 
webdriver-manager
seaborn
nltk
wordcloud
plotly
streamlit
redis
websocket-client
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
httpx>=0.24.0
